This repo is intended for a package where all the bits and pieces that are running ML on our field cameras can be gathered and organized sensibly.
The short-term motivation is to switch scrubcam over to using this
package and get this package onto the PyPI to simplify setup of
scrubcams.

This first commit includes pulling the current code for running models
on the Google Coral USB accelerator out of the dnntools package and
slightly refactoring it.