Switch ImageClassifierHandler over to pycoral
Switched this class over to use pycoral instead of edgetpu library.

Note: object detection related classes and functions are still here
dependent on edgetpu until they are also converted BUT because the
pycoral apt package and edgetpu apt package de-install each other,
this module is not currently runnable. To test it, I was commenting
out those classes and functions but didn't want to commit with those
commented-out bits as it makes the diff not tight.