Better organize where created models are put
And also where other associated files generated in the pipeline are
put. Updated README tutorial to reflect change.  Mostly the intent is
to keep them out of the repository where they would clutter things up
but also means that tflite models, derived edgetpu tflite models, and
associated label maps are neatly in a single folder that is a
subfolder of "tflite_train/" (assuming user follows pattern of the
README).