Update detection bits to not require pycoral
They now use tflite-runtime directly using elements from the pycoral
code that are directly added to this repo.  This was prompted by a
need to use `camml` on contemporary versions of Raspberry Pi OS that
default to versions of Python that `pycoral` itself has not been
updated to work with.